Move lifetime tests to their own file.
